pyproject.toml
[tool.poetry]
name = "ai-girlfried"
version = "0.1.0"
description = ""
authors = ["Marcus <marcus@pyteam.pro>"]
readme = "README.md"
packages = [{include = "ai_girlfried"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.238"
openai = "^0.27.8"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
playsound = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
